Improve Conv hashing, remove infinite loop in top-level graph splitting
