+++ include/flexflow/fftype.h
@@ -1,22 +1,21 @@
-#pragma once
+#ifndef _FF_TYPE_H
+#define _FF_TYPE_H
+
+#include <cstddef>
 #include "flexflow/ffconst.h"
 
 namespace FlexFlow {
 
 class LayerID {
 public:
-  LayerID() : id(0) {}
-  LayerID(size_t _id) : id(_id) {
-    assert(is_valid_id());
-  }
-  bool is_valid_id() const {
-    return (id >= LAYER_GUID_FIRST_VALID && id <= LAYER_GUID_LAST_VALID);
-  }
-  bool operator == (const LayerID & other) const {
-    return this->id == other.id;
-  }
+  LayerID();
+  LayerID(size_t id);
+  bool is_valid_id() const;
+  friend bool operator==(LayerID const &lhs, LayerID const &rhs);
 public:
   size_t id;
 };
 
 };
+
+#endif // _FF_TYPE_H
