+++ include/flexflow/utils/dot/record_formatter.h
@@ -1,0 +1,17 @@
+#ifndef _RECORD_FORMATTER_H
+#define _RECORD_FORMATTER_H
+
+#include <vector>
+#include <sstream>
+
+class RecordFormatter {
+  friend RecordFormatter &operator<<(RecordFormatter &r, std::string const &tok);
+  friend RecordFormatter &operator<<(RecordFormatter &r, int tok);
+  friend RecordFormatter &operator<<(RecordFormatter &r, RecordFormatter const &sub_r);
+  friend RecordFormatter &operator<<(RecordFormatter &r, std::ostringstream &oss);
+  friend std::ostream &operator<<(std::ostream &s, RecordFormatter const &r);
+private:
+  std::vector<std::string> pieces;
+};
+
+#endif // _RECORD_FORMATTER_H
