+++ scripts/osdi22ae/resnext-50.sh
@@ -1,5 +1,5 @@
 echo "Running ResNeXt-50 with a parallelization strategy discovered by Unity"
-$FF_HOME/build/examples/cpp/resnext50/resnext50 -ll:gpu 4 -ll:fsize 14000 -ll:zsize 14000 -b 16 --budget 20
+$FF_HOME/build/examples/cpp/resnext50/resnext50 -ll:gpu 1 -ll:fsize 14000 -ll:zsize 14000 -b 16 --budget 20 --search-num-nodes 1 --search-num-gpus 4
 
-echo "Running ResNeXt-50 with data parallelism"
-$FF_HOME/build/examples/cpp/resnext50/resnext50 -ll:gpu 4 -ll:fsize 14000 -ll:zsize 14000 -b 16 --budget 20 --only-data-parallel
+#echo "Running ResNeXt-50 with data parallelism"
+#$FF_HOME/build/examples/cpp/resnext50/resnext50 -ll:gpu 4 -ll:fsize 14000 -ll:zsize 14000 -b 16 --budget 20 --only-data-parallel
