+++ tests/unit/test_dot.cc
@@ -1,5 +1,5 @@
 #include "gtest/gtest.h"
-#include "flexflow/utils/dot_file.h"
+#include "flexflow/utils/dot/record_formatter.h"
 
 TEST(record_formatters, basic) {
   RecordFormatter rf, rf2, rf3;
